refactor: 크루가 보는 시트 페이지 분리
