+++ front/src/App.tsx
@@ -6,7 +6,8 @@ import SelectUser from '@pages/SelectUser';
 import Schedule from '@pages/Schedule';
 import Coach from '@pages/Coach';
 import NotFound from '@pages/NotFound';
-import Sheet from '@pages/Sheet';
+import CrewSheet from '@pages/CrewSheet';
+import CoachSheet from '@pages/CoachSheet';
 import CrewHistory from '@pages/CrewHistory';
 import Header from '@components/Header';
 import ScheduleProvider from '@context/ScheduleProvider';
@@ -22,7 +23,8 @@ const App = () => {
         <Route path={`${ROUTES.COACH}/:id`} element={<Coach />} />
         <Route path={`${ROUTES.RESERVATION}/:id`} element={<Reservation />} />
         <Route path={`${ROUTES.SCHEDULE}/:id`} element={<Schedule />} />
-        <Route path={`${ROUTES.SHEET}/:id`} element={<Sheet />} />
+        <Route path={`${ROUTES.CREW_SHEET}/:id`} element={<CrewSheet />} />
+        <Route path={`${ROUTES.COACH_SHEET}/:id`} element={<CoachSheet />} />
         <Route path={`${ROUTES.HISTORY}/:id`} element={<CrewHistory />} />
         <Route path="/*" element={<NotFound />} />
       </Routes>
