+++ front/src/components/ReservationInfo/index.tsx
@@ -3,16 +3,19 @@ import * as S from './styles';
 
 import ScheduleIcon from '@assets/schedule.svg';
 import ClockIcon from '@assets/clock.svg';
+import LeftArrowIcon from '@assets/left-arrow-disabled.svg';
 
 interface ReservationInfoProps {
   image?: string;
   name?: string;
   dateTime?: string;
+  isView: boolean;
+  onClick: () => void;
 }
 
-const ReservationInfo = ({ image, name, dateTime }: ReservationInfoProps) => {
+const ReservationInfo = ({ image, name, dateTime, isView, onClick }: ReservationInfoProps) => {
   return (
-    <>
+    <S.InfoContainer>
       <S.CoachImg src={image} alt="코치 프로필 이미지" />
       <p>{name}</p>
       <S.DateWrapper>
@@ -23,7 +26,8 @@ const ReservationInfo = ({ image, name, dateTime }: ReservationInfoProps) => {
         <img src={ClockIcon} alt="시계 아이콘" />
         <span>{getHourMinutes(dateTime as string)}</span>
       </S.DateWrapper>
-    </>
+      {isView && <S.ArrowIcon src={LeftArrowIcon} alt="화살표 아이콘" onClick={onClick} />}
+    </S.InfoContainer>
   );
 };
