+++ front/src/components/Sheet/styles.ts
@@ -1,35 +1,5 @@
 import styled from 'styled-components';
 
-const InfoContainer = styled.div`
-  position: relative;
-  width: 30%;
-  margin-right: 50px;
-  border-right: 1px solid ${({ theme }) => theme.colors.GRAY_200};
-  font-weight: bold;
-
-  > p {
-    font-size: 20px;
-  }
-`;
-
-const ArrowIcon = styled.img`
-  position: absolute;
-  left: -1em;
-  bottom: 1em;
-  width: 50px;
-  height: 50px;
-  padding: 10px;
-  border-radius: 50%;
-  background-color: ${({ theme }) => theme.colors.BLUE_800};
-  cursor: pointer;
-
-  :hover {
-    opacity: 0.7;
-    transform: scale(1.1);
-    transition: ease-in-out 0.2s;
-  }
-`;
-
 const SheetContainer = styled.div`
   width: 100%;
   overflow: scroll;
@@ -67,4 +37,4 @@ const SecondButton = styled.button`
   border: none;
 `;
 
-export { InfoContainer, SheetContainer, ButtonContainer, FirstButton, SecondButton, ArrowIcon };
+export { SheetContainer, ButtonContainer, FirstButton, SecondButton };
